+++ desktop-widgets/simplewidgets.cpp
@@ -18,7 +18,7 @@
 #include "core/qthelper.h"
 #include "libdivecomputer/parser.h"
 #include "desktop-widgets/divelistview.h"
-#include "core/display.h"
+#include "core/selection.h"
 #include "profile-widget/profilewidget2.h"
 #include "commands/command.h"
 #include "core/metadata.h"
