+++ desktop-widgets/tab-widgets/TabDiveStatistics.cpp
@@ -3,7 +3,7 @@
 #include "ui_TabDiveStatistics.h"
 
 #include <core/qthelper.h>
-#include <core/display.h>
+#include <core/selection.h>
 #include <core/statistics.h>
 
 TabDiveStatistics::TabDiveStatistics(QWidget *parent) : TabBase(parent), ui(new Ui::TabDiveStatistics())
